Refactoring code for the first phase
=== Details ====
* Organize code
* Create custom exceptions
* Add unit tests
* Fix some maven dependencies
* Add configuration file for logs
