--- src/main/kotlin/com/docletinterface/doclet/logic/JsonCommandParser.kt
@@ -1,12 +1,0 @@
-package com.docletinterface.doclet.logic
-
-
-class JsonCommandParser {
-
-    fun parseObject(json: String, classRef: String): Any {
-        //val classInit = Class.forName(classRef)
-        //return ObjectMapper().readValue(json, classInit)
-        return ""
-    }
-
-}
