+++ src/main/kotlin/com/docletinterface/ui/domain/Request.kt
@@ -1,4 +1,4 @@
-package com.docletinterface.doclet.domain
+package com.docletinterface.ui.domain
 
 open class Request{
